#19 user cant spend negative tokens

+++ trndiiapp/app/Http/Controllers/TransactionsController.php
@@ -139,13 +139,13 @@ class TransactionsController extends Controller
         if($stripeId != ''){
 
             $nbTokensSpent = $request->input('Tokens_To_Spend');
-            if($request->has('Tokens_To_Spend')&&$user->tokens>=$nbTokensSpent){
+            if($request->has('Tokens_To_Spend')&&$user->tokens>=$nbTokensSpent&&$nbTokensSpent>0){
 
                 $this->itemRepo->addTotalTokens($nbTokensSpent,$id);
                 $this->userRepo->removeTokens($user,$nbTokensSpent);
             }
             else{
-                return back()->with('error', 'You do not have enough tokens to spend that amount');
+                return back()->with('error', 'You cannot spend the amount of tokens you entered');
             }
 
             app('App\Http\Controllers\ItemsController')->numTransactions($id);
